requirements.txt
google-cloud-bigquery
google-cloud
google-cloud-storage

pandas